requirements.txt
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
construct==2.10.68
construct-typing==0.6.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonalias==0.1.1
python-telegram-bot==22.5
requests==2.32.5
sniffio==1.3.1
solana==0.36.9
solders==0.26.0
typing_extensions==4.15.0
urllib3==2.5.0
websockets==12.0
